requirements.txt
pymupdf>=1.24
anthropic>=0.40
pydantic>=2.7
python-dotenv>=1.0
fastapi>=0.115
uvicorn[standard]>=0.30
redis>=5.0
websockets>=12.0
python-multipart>=0.0.9
httpx>=0.27
